[Update] Handle warnings
Major:
- To disable warnings about unsafe functions such as strcpy, fopen, etc. --> #define _CRT_SECURE_NO_WARNINGS
- Some float/double conversion using the float versions of the functions
- Fix potential bug with a function possibly not returning any value.
- Some std::size_t --> int conversions

+++ ECell_Engine/src/Solver/ODE/GeneralizedExplicitRK.cpp
@@ -116,9 +116,9 @@ void ECellEngine::Solvers::ODE::GeneralizedExplicitRK::Initialize(const ECellEng
 	delete[] ynp12;
 	delete[] yn_ext;
 
-	systemSize = system.size();
+	systemSize = (unsigned short)system.size();
 
-	extEqSize = dataState.GetEquations().size();
+	extEqSize = (unsigned short)dataState.GetEquations().size();
 	externalEquations.reserve(extEqSize);
 	for (auto [equationName, equation] : dataState.GetEquations())
 	{
